[driver_test_realm] Migrate to dtr_exposes/offers
Migrates all users of the deprecated offers and exposes fields to the
new dtr_exposes and dtr_offers. The new field allows for the custom type
to be removed and uses the fuchsia.component.testing capability type. It
also provides a direct capability route internally instead of manually
forwarding the service directory for exposes.

Change-Id: I55833e09fcb34187c19ceb0a7f6c88d2e05a7c10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112694
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>