[iptables] Use iptables-restore to make multiple policy changes
iptables-restore takes input from stdin in the form of table
specifications. There is little documentation about its exact format,
except that it is what iptables-save outputs.

Modify the test so that iptables-restore is ran in the realm's
collection (added in tqr/881532), with a handle to its stdin.

Bug: 307908515
Change-Id: I9ae8366cf201c9cea5a104781d2b86a91db4ffa2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114559
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: David Song <wintermelons@google.com>