[bazel][clang] Allow user-defined sysroot filegroups and paths
This is needed to migrate in-tree IDK generation from GN to Bazel, since
the migration will change some libraries from the sysroot to be built
with Bazel instead of GN, and a custom target will likely be needed to
generate the sysroot, which means we can no longer hardcode the paths to
Fuchsia sysroot in fuchsia_clang.

Default values are provided for newly add attributes to match existing
behavior.

Change-Id: I71b0cc9a40ca5f5f52db63d936e31433decd10ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115432
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Jay Zhuang <jayzhuang@google.com>