[sestarnix] Use a track_permission_stub macro for unenforced checks
Introduce a macro akin to track_stub!(), for permission checks that
are not currently enforcing, to ensure that they are not forgotten
about.  Each permission check "stub" must start with a TODO() which
will be passed to track_stub!() to allow tracking of the number of
implemented-but-ignored checks.

Change-Id: I4dc9ae2c0e1b356bee5495b969b48ca6ac1e6588
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114392
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>