[starnix] Reflect reserved brk space in mappings
Starnix reserves some space after brk so that the addresses from that
range are not selected and brk can be safely grown.

In preparation for switching to Starnix-based address selection, change
the logic so that this reserved space is tracked by Starnix (and not by
Zircon only as it was before).

Additionally, this makes this behavior consistent across forks, as
previously we were not replicating the reservation.

Bug: 309452853, 365834029, 333111647
Change-Id: I2116f7293be9d11bfd48cb3b6c254227d7754bbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116297
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Maria Glukhova <mariagl@google.com>