[usb][adb] take peripheral device offline when one function clears its interface
When `SetInterface` of any one function is reset, take the peripheral
device offline. When all functions have called `SetInterface` with a
valid interface, bring the device online.

Take the device offline when `Stop` is called in adb and bring it online
when `Start` is called.

UsbDciInterfaceServer in UsbPeripheral must live on a different dispatcher
to prevent blocking itself on callback.

Bug: b/357604844

Test: tested on Vim3
      usb-adb-unittest
      usb-peripheral-unittest
      usb-virtual-bus-test

Change-Id: I1c1b0d1c6987233c419141c2ff50bddc29790085
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116695
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>