[tools][artifactory] Only upload artifacts related to the last build.
This is a reland of [1], after the fix in [2], for a bug that justified
the revert.

Ensure that `artifactory` only tries to upload Ninja artifacts that
correspond to the last successful build. This avois uploading stale
artifacts from previous incremental builds, or artifactory
complaining that it does not find breakpad symbol files matching
(stale) debug files in the build directory.

- Add new `lastBuildOnly` argument to NewBuildAPIClient()
  function is added, which creates a BuildAPIClient
  instance to return filtered results.

- Add `NewModulesForLastBuildOnly()` function to return
  a `Modules` instance corresponding to these filtered
  modules.

- Use `NewModulesForLastBuildOnly()` in the artifactory
  `up` command implementation.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1113733
[2] https://fuchsia-review.googlesource.com/c/fuchsia/+/1114412

Bug: 42175326
Change-Id: I13548d62b85f0f9a971f32a134c91fbf44078cee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114414
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: David Fang <fangism@google.com>