[ffx] Remove redundant ssh stderr read.
If there's an error setting up SSH and running a compatibility check,
then there is no need to read the stderr a second time in the midst of
encountering an error. This will remove the original failure message,
causing the error reported to the user to be an empty string, which is
reported as "unknown" and then tells the user to report an issue, while
hiding the original error.

Bug: 362976369
Change-Id: I0f386a9aabeb8356dbbc24c301883e684226e820
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116852
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>