[assembly] default assembled_system() to use eng platform AIBs
All in-tree uses of assembled_system() are for eng build-type
assemblies, so make that the default to reduce the common user
error of not setting this option correctly.

Bazel assemblies are left as-is, because they can't accidentally
use the eng build-type AIBs when not setting the flag correctly,
as their action sandboxes won't include the eng build-type AIBs
at all.

Bug: b/365667289
Change-Id: I57fe3de3596759d7c65266b2954e47024615b422
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116873
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>