[bus][pci] Attempt to use INTA if InterruptPin reads 0x00
According to the PCI spec, if the Interrupt Pin register in the
configuration header reads 0x00 it means the device function does not
have an interrupt. Crosvm does not appear to set this pin correctly
presently, so for now we'll try pin 1 (INTA) for the purposes of
checking the interrupt routing table we've built. This should not cause
a problem on other platforms because the interrupt configuration is
hardcoded, or does not have routing information for that device and pin
combo if InterruptPin was zero. In fact, it's generally rare to not have
a legacy interrupt pin at all.

Bug: b/365837900
Change-Id: I82c645c93c6f922c773cf5fc404d0b09eb326dc3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116894
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>