[rfc] Monotonic clock suspension and the boot timeline
Adds an RFC to further elaborate on the montonic clock suspension
described in RFC-0230. This RFC also introduces a new timeline, called
the boot timeline, which programs that need to be aware of time spent
suspended can switch to before we pause the clock.

This is a high level RFC intended to describe the conceptual changes to
the system and the areas of the system that need to be modified before
we can pause the clock. More concrete designs and implementation details
will follow in future RFCs or design documents.

Bug: 341785588, 328306129
Change-Id: Ie1eef64bc8adb60c190e072f2272bec008c2d7d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102437
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Thomas Gales <tgales@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>