[ffx] Use absolute paths in isolate dirs
When the daemon is autostarted, there is no guarantee it starts from
the same directory as the invoking CLI. This means that if an isolate
dir is passed in as a relative path (e.g. `ffx --isolate-dir foo ...`)
the daemon may end up using a different isolate dir. The user-visible
behavior is that CLI cannot make a connection to the daemon (because the
ascendd socket ends up in a different location).

To fix this, when a isolate dir is specified as a relative path, we
canonicalize it into an absolute path.

Fixed: 362704093
Change-Id: I8e8b2983c905481cf8ed13355449a7727cfd5018
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116679
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>