[ffx] when getting config objects, merge levels
The default behavior of `ffx config get` is to stop when it reaches
the first level with an existing value. Unfortunately, if the
value in question is an "object", then that behavior is not
helpful. E.g. if at the user level, we have `obj: { foo=bar }`
and at the run-time level we have `obj: { x=y }`, then when querying
`obj.foo` or obj.x` we get the respective values. But if we just
query `obj`, we only get `obj.x` (since that is the first one found).

Instead, when we find an object, we should merge in from all the
levels, ignoring keys that already exist, to match the semantics
of individual field queries.

Fixed: 359934352
Change-Id: I47c201aa30ef165b104fae166a36078f0fb6d126
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117175
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>