[starnix] Add channel proxy call to Starnix manager
This API will be used to allow the Starnix runner to proxy channels on
behalf of the Starnix kernel. This will allow messages on the proxied
channels to wake the kernel when it is suspended.

Change-Id: I355d17040bbb01aa675554a50842d5a38b9b4df9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115415
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
API-Review: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>