[gn] Enable no_stamp_files option.
Now that fxr/1115414 has been submitted, and the corresponding
Ninja binary rolled into the Fuchsia tree, the new GN
`no_stamp_files` option in `//.gn` can be turned on.

Its effect is to replace stamp files in the Ninja build plan
(which are actual files written to disk during the build)
with phony aliases (which are paths like `phony/lib/foo`
which are never written to disk).

This is necessary because GN is going to remove support for
stamp files entirely in the near future, as requested by the
Chromium team for improve their Siso tool.

Fixed: 363930634
Change-Id: Iab866c0b4ecaa7f3621901d5b12aa08f4e8473b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115537
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>