[fidl] Rename fuchsia.device.manager to fuchsia.system.state
The old library was based on an older name for the fuchsia driver
manager. However these days the only protocol that remains in it is
served by the shutdown-shim component with clients that are not driver
manager. Renaming it to more accurately relfect what it is should help
reduce any confusion with respect to it's purpose.

Change-Id: I2247c246735826c0804e758a205b581d93263ebc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115616
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>