[starnix] Implement ProxyWakeChannel in runner
This implements the new ProxyWakeChannel in the Starnix runner. The
runner creates a new thread which waits for either of the proxy
endpoints to become readable (or receive a peer closed). When a message
is available on one endpoint it is proxied to the other.

If a message is proxied into the container channel endpoint,
EVENT_SIGNALED is raised on the event provided to the ProxyWakeChannel
call. The will wake the container if it is suspended. Future suspend
attempts will fail until the client of the ProxyWakeChannel call clears
the signal from the event.

Change-Id: Ied8d51c3e901c025c85fb1a3ea92ccb731f13356
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117712
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>