[physload] Test harness will own the kernel storage.
The API for a physload handoff test consisted in handing off the kernel
storage to the test. This is problematic because, when the test routine
returns, the kernel storage is released (the allocation member is
destroyed). Then the test harness prints the routine, and follows up
with calling abort.

Why is this problematic? Well any interaction that may allocate memory
after the test completes, may allocate over the image of the currently
executing image.

This introduced a source of flakyness for this tests, because depending
on how long it took to dump the stack (in this case), it would fire a
tefmo check against failed assertiongs.

Fixed: 366078334
Change-Id: Ie7e2c79a09dc503573c17dc0be33d53c24ccc3b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117532
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>