[assembly] Give developer overrides allowlisting a better error message
The definition of developer overrides targets is currently restricted
to local-only locations (they cannot be checked into the tree and used
from there).

This is to surface use-cases and discuss whether they are best
defined as local settings, or features that need to be added to
the platform configuration schema and rules.

This change gives this restriction a better error message than
the 'dependency not allowed' visibility restriction that it
currently has.

Bug: b/366083446

Change-Id: I2957aa5188951864d24f1615231abfa88263367f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117776
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>