[ffx][emu] Add context to 'persistent emulator'
Due to the ordering of the subcommands in the ffx documentation ("start"
comes before "stop"), at the point of reading through the `reuse`
switches it is unclear what a persistent editor means. This adds a key
piece of helpful context for the reader in the documentation of the
"start" subcommand.

Change-Id: I80a65b4d68673c25bbaac42ac40ccef5ac1488a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116493
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Thomas Zander <thomasalpinus@google.com>