[sestarnix] Fix SeLinuxApi<> seek handling for reads & writes
The "selinuxfs" has a number of files implementing administrative
and query APIs, which are seekable, but have asymmetric behaviour,
with `read()` operations behaving as for a normal file, but
`write()` operations not updating the seek position, and varying
in whether they ignore or error if made from a non-zero seek
position.

Most fundamentally, this fixes the implementation to correctly
return end-of-file, rather than mistakenly returning the same
content over and over again.

Documentation of the expected behaviour of the SeLinuxApi<> helper
is updated to reflect the behaviours observed with the Linux
implementation.

Bug: 365812498, 304775883
Change-Id: Ifb30bacc18bf2daaf56e10c8dfc997023416b12b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116496
Reviewed-by: Ali Zhang <alizhang@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>