[wlan][fidl] Make FIDL libraries versioned
Add `@available(added=HEAD)` to libraries in the "fuchsia" namespace
that are included in an SDK [category] but do not currently specify
`@available`. This will allow the `fidl()` template to verify all
libraries in the "fuchsia" namespace in a subsequent CL.

In one library, `7` is replaced with `HEAD`. The `7` was added in
https://fxrev.dev/664072 but not removed when the category was fixed in
https://fxrev.dev/679350.

Bug: 364628320
Change-Id: Ibb90db74ac9f53985db22294970d39684c0d24d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114043
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>