[time] simplify some code using `Rc<Cell<_>>`
A recent CL introduced some shared state, and managed that
state using `Arc<Mutex<_>>` (with a wrapper to make that
deadlock-free).

This works, but the code to manage requires the reader to
convince themself that the potential panic will never happen.

In this CL, we switch to using `Rc<Cell<_>>`, since the `get()`
and `set()` operations on a `Cell` can never deadlock or panic.

To make this possible, we modify some of the tests to avoid
the use of `BoxFuture`. The issue with `BoxFuture` is that
it requires the underlying future to be `Send`.

By design, `Rc` and `Cell` are not `Send`, so we cannot
use `BoxFuture` with a future that captures an `Rc<Cell<_>>`

Fortunately, we're able to mechanically replaced the calls
to `boxed()` with calls to `pin!()` the relevant futures
instead.

Change-Id: I2641ff362aa7511dda2d8030021a46e7f107a8e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118064
Reviewed-by: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>