[dtr] Use test_package field instead of pkg
The new test_package field is now part of the partner sdk and supports
both the subpackage workflow as well as the old workflow where the
driver is in the test package directly.

This also adds a new field to the start args called
boot_driver_components which can be used to decide if a driver should be
a boot driver, aka started in the boot collection which has more
capabilities available to it.

Change-Id: I439c5cf5aa25e86822be041e4c34a56ed3fe410d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113381
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>