[trace_processing] revise computation of suspend metrics
In some conditions, it's possible for the trace to contain
a slice that began before the test itself.

When such a slice is present, the previous method of computing
`total_time` would misrepresent the duration of the test,
and thereby cause the `UnsuspendedTime` to be high, and the
`SuspendPercentage` to be low.

To resolve this, we compute the trace start time based on the
first scheduler slice. That slice should be generated very
soon after the test enables tracing. So it's a good
representation of when the test started.

Bug: 363067753
Test: manually verified that a trace from a recent tryjob achieves
Test: ~60% suspend time.
Change-Id: I03d8360157a1ae18204ec7c3b9d84b9d9942cc0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117574
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>