[fidl] Correct added=7 for libraries in internal SDK category
Since FIDL libraries in the "internal" SDK category are not in an SDK or
stable, they should be `@available(added=HEAD)`.

`@available(added=7)` was added in https://fxrev.dev/664072 but not
removed when the category was fixed in https://fxrev.dev/679350.

Bug: 364628320
Change-Id: Ie8379d74fd7bf28cae7fd57b94490d485adabf47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114044
Commit-Queue: David Dorwin <ddorwin@google.com>
API-Review: Marie Janssen <jamuraa@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>