[intl][fidl] Make generated string FIDL libraries versioned
Add `@available(added=HEAD)` to generated string libraries. This is
required in order for the `fuchsia.intl.l10n` FIDL library to pass
FIDL platform versioning verification, which will be enabled in the
`fidl()` template for all libraries in the "fuchsia" namespace in a
subsequent CL. Previously, the FIDL content appeared to be "versioned
under platform 'unversioned'."

This CL assumes that all such string libraries are unstable.

Bug: 364628320
Change-Id: I32dd3cde24a5b5d4f0c843a60f243358becb8fd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114045
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>