[rust-3p] Re-land "Update rustls to 0.21.7""
This patch re-lands the rustls update, which was reverted in
http://fxrev.dev/1089232, or 04c91aa757acc91cd6a83b1216f7460db9c3f422.

The issue we ran into was in `src/lib/fuchsia-hyper/src/not_fuchsia.rs`.
The old version would error out if `rustl_native_certs` failed to parse
any certificates, but didn't actually error out if it found no
certificates to try to parse. The new version treated this as an error.

The fix here then is to first check that we actually have certificates
to add before checking if we ran into any problems trying to parse them.

Note that this does not remove rustls 0.19.7, we'll remove that in a
follow on CL when this sticks.

Bug: 364892934

Change-Id: Icdac0f83c2cb2a8d791ba073a2bd8f9df92b2add
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112055
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>