[ffx] Use ffx_config::get_host_tool() for plugins
The `ffx_config::get_host_tool()` function was introduced
in [1] to allow command-line overrides for the SDK
host tools (e.g. with `ffx --config sdk.overrides.cmc=<path>`),
however only the `emulator` ffx plugin used it.

This CL ensures that other plugins are using the same
function for a consistent experience. Without it, the
`ffx assembly` plugin would ignore the overrides and fail
to find the right host `cmc` tool in [2].

This change is required by the associated bug because in
the future, SDK host tools used in-tree will be built
directly by Bazel on-demand. As a consequence, their manifests
will point to Bazel targets (e.g. `@fuchsia_idk//tools/x64:cmc`)
instead of file paths, and the actual file path will be passed
using `sdk.overrides.<name>` by the Bazel action invoking the
tool (that path cannot be predicted in advance and written to
the manifest, hence the need for an override).

[1]: https://fuchsia-review.googlesource.com/c/fuchsia/+/994926
[2]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115539

Bug: 367274537
Change-Id: I863b413252fa0ffd1d5a3e9d98105e552ac95902
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119632
Reviewed-by: Brian Bosak <bbosak@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>