[bazel_sdk] Override host tool paths when invoking ffx.
This is a reland of [1], which should work thanks to [2].

Ensure that `ffx` and its plugins do not rely on the SDK manifest
to locate the host tool they may invoke at runtime, by passing their
location directly on the command line.

New functions are added to `ffx_tool.bzl` to retrieve the start
of the command-line invocation, that includes the necessary
`--config "sdk.override.<tool1>=<path1>,sdk.override.<tool2>=<path2>"`
argument.

This is required to allow moving the generation of the IDK to the
Bazel workspace. In this scenario, host tools will be built on-demand
as Bazel targets, and their location cannot be predicted before the
build, making it difficult to write them to the SDK manifest in advance.

(The SDK manifest will instead contain Bazel labels such as
`@fuchsia_idk//tools/x64:cmc` that `ffx` definitely cannot convert to
a proper host path on its own).

[1]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114415
[2]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119632

Bug: 367274537
Change-Id: If8da9687d7b873a942860540a98f72d4871d87f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115539
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>