[ffx] extend ssh timeout
We are seeing ssh timeouts when connecting to emulators
(see https://b.corp.google.com/issues/364345273#comment6).
The delays are probably due to a Cavium issue (b/343550891)
in which a process (e.g. an emulator) can be delayed for
up to 15 seconds.

This change mitigates the problem, while the root cause is being
addressed.

(This change also improves an error message -- the previous version
resulted in a lot of lost time due to confusion about the error in
question.)

Bug: 364345273
Fixed: 366582564
Change-Id: I24378d712a5ad01978f7371130cd1c82399df0c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119775
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>