[ffx][emu] Moving eprintln statements to debug!
This changes some of the more verbose messages that were printed to
stderr when running `ffx emu start --verbose` to be in the debug log
instead.

This messages were too verbose to be useful by the user, and not very
actionable.

The command line generated is still printed to stderr to enhance the
logging of running the emulator in infra.

Change-Id: If63ee47ab556fbabd8022b3e1e466abb8999b3d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119756
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>