[build][fidl] Enforce versioning requirements for in-tree libraries
Require that in-tree FIDL libraries fall into one of the following:
* Stable, included in an SDK, requires compatibility tests, and
  versioned at "fuchsia"
* Unstable, included in an SDK, requires compatibility tests, and
  versioned at "fuchsia:HEAD"
* Unstable, not included in an SDK, does not require compatibility
  tests, and versioned at "fuchsia:HEAD"
* In the "internal" SDK category
* Unstable, not included in an SDK, does not require compatibility
  tests, and "unversioned"
  tests or not included in an SDK, and versioned at "fuchsia:HEAD"
* Vendor SDK libraries that are included in the "partner" SDK category
  but unstable, do not require compatibility tests, and are
  "unversioned"
* Test-only, unstable, not included in an SDK, does not require
  compatibility tests, and "unversioned"

There are some existing exceptions to these about requirements that have
been identified with TODOs to resolve the issues.

There are three libraries that are unstable but cannot be made
versioned at "fuchsia:HEAD" because they are used by SDK prebuilts. The
internal "zx" library also does not match one of these cases and cannot
be correctly marked versioned because it is used by ulib in the SDK. The
Banjo `fuchsia.sysmem` library is versioned but not in an SDK category.
The `fdf` library  is in the "experimental" SDK category but
unversioned.

In addition, there are 13 currently unversioned libraries on which CTF
tests depend.

Bug: 325669391
Bug: 330947474, 364293789, 364294648, 306258166, 364422340, 364727245
Bug: 365558387, 365602422, 365602428
Change-Id: I1da9b1e9a838fdf2fd52a5efa51abe4500ea0a01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114154
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>