[sestarnix] Eagerly initialize FileSystem labeling details
Previously labeling information for FileSystems was resolved on-
demand, to simplify handling of the pre-policy-load state.

FileSystems are now eagerly labeled, either at construction, or
when a policy is first loaded. This requires splitting the
FileSystem security state initialization across several hooks, each
with dependencies that can be accommodated during Starnix boot:

1. Initialization of SELinux mount-options, based on the FileSystem
   mount options.
2. Resolution of the FileSystem labeling information, based on the
   loaded policy, and SELinux mount options for the FileSystem.

The kernel-global SELinux state is extended to accommodate a queue
of FileSystems used to track those constructed prior to policy load
so that the SELinux subsystem can enumerate and resolve labels for
them as soon as a policy is first loaded.

Bug: 366405587, 366405530, 367309046
Change-Id: I70e27e91f2e1ac1ef35cd0b82ccd2bbaaf226416
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117992
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nelly Vouzoukidou <nellyv@google.com>