[tee] Add system properties to TA manager
Adds a base system properties json file to TA manager. In an actual
deployment, this file is expected to be combined with data from the
build and other sources.

Adds test to verify that the system properties hosted by TA manager can
be routed to other components and successfully read. When the Property
APIs are implemented, the test will expand to TAs in addition to test
components.

Updates tee_properties lib to use serde_json5 to support comments in the
system properties file.

Bug: b/361406102
Test: fx test //src/tee/manager
Change-Id: I2d83367a96563ef280cc45461153486b7e876ce3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114213
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>