[bazel_sdk] Remove @fuchsia_sdk_common mentions from shared rules.
This is the first in a series of two CLs which refactor how shared
definitions between SDK rules and the in-tree Bazel workspace
are exposed and used.

The `//build/bazel_sdk/bazel_rules_fuchsia/common` directory
contains shared definitions used by both the SDK Bazel rules and
the in-tree Bazel workspace.

When used in-tree, the directory is currently wrapped as an
external repository named `@fuchsia_sdk_common`, however this is
not required at all for OOT workspaces, which do not define it.

This CL removes a few hard-coded references to `@fuchsia_sdk_common`
for some of the functions defined in the directory, making these
usable from `@fuchsia_sdk` and `@fuchsia_clang` as well.

This is done by introducing a new `common_package_prefix` argument
to several of the functions, and adjusting only two call sites
outside of the `common` directory.

Bug: 366156651
Change-Id: Ia3a383863ced1d2c72d86ad68192cc4421bc11af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120392
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>