[bazel_sdk] Change how common definitions are shared between workspaces.
This is the second CL in a series that refactors how the shared
definitions of the Bazel SDK rules and in-tree workspace are used.

This CL changes how `//build/bazel_sdk/bazel_rules_fuchsia/common`
is exposed to the SDK and in-tree Bazel workspaces. This is done to
simplify fuchsia_clang_repository() which can now only create a
single top-level `common` symlink to that directory, as the whole
content under it can be used safely now.

To do so, a new convention is enforced, where any external repository
which wants to use these shared definitions should reference them
through a top-level `common` package/directory.

For `@fuchsia_sdk//common`, this is already the case, as this maps
directly to `//build/bazel_sdk/bazel_rules_fuchsia/common` by
construction.

For `@fuchsia_clang//common`, this is done by creating a simple
symlinks to the same location, which simplifies the
`fuchsia_clang_repository()` rule, which no longer needs to keep
track of the exact set of files to symlink.

For `@fuchsia_sdk_common`, which is only used by the in-tree
build (and the Bazel SDK test suite which runs in-tree), a new
`build/bazel/local_repositories/fuchsia_sdk_common` directory
is created with a single `common` symlink as well.

This setup allows removing the `common_package_prefix` argument
introduced in a previous CL, as now all functions defined in
the common area can assume they can reference other files from
the same set using `//common:<path>` or
`//common/<package>:<path>`.

This simplification will also allow moving more definitions
to the common area in the future. In particular those related
to C++ toolchain definitions.

Bug: 366156651
Change-Id: I56bd74ba303cec8acd8042f995e186428d85f969
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120393
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>