[tee] Add TA properties template
This template properties file will be included with each TA, with the
option to override it. A future change will add validation, which the
template will fail and therefore require TAs to provide a valid property
file.

This change also adds a read_props test TA which currently checks for the
existence of the properties file. This TA can be extended to exercise
each internal core GetProperty API.

Bug: b/361406102
Test: fx test //src/tee/manager
Change-Id: I856149ff3b6526af56b585d2f9c57c90b0bd666c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118496
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>