[superproject] Roll llvm-project/libc [Clang] Do not implicitly link C libraries for the GPU targets (#109052)
Summary:
I initially thought that it would be convenient to automatically link
these libraries like they are for standard C/C++ targets. However, this
created issues when trying to use C++ as a GPU target. This patch moves
the logic to now implicitly pass it as part of the offloading toolchain
instead, if found. This means that the user needs to set the target
toolchain for the link job for automatic detection, but can still be
done manually via `-Xoffload-linker -lc`.
GitOrigin-RevId: 3c6214e4d30b656d4bfcb0d87f9f436ab3f20ba4
Original-Revision: 19a8a61235677da7a05b722f5d117f4b141d8de8
Roller-URL: https://ci.chromium.org/b/8736443758180567553
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I03a58cdaa0a1e50db3de4ee65ce9c41b976f3d00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121353