[starnix][sysfs] Store KObject in device registry
We now store the KObject for each registered device. Storing this object
will let us implement /proc/misc and /proc/devices in a future CL.

In order to store the KObject for a device, we need to add the device to
the KernelObjects store before registering the device, which is the
opposite order from what we did previously.

This CL also cleans up the implementation of the device registry to use
more type names and to avoid storing the DeviceMode redundantly inside
each of the MajorDeviceRegistry objects.

Change-Id: I70deb083f7223b4fa69add384995c9083f622bdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120213
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>