[rust-3p] Add zerocopy 0.0.0-alpha.0
This is identical to 0.8.0-alpha.21, and was published explicitly to
support transitioning to that version. Our current vendored version is
0.8.0-alpha.5. If we tried to add 0.8.0-alpha.21, Cargo would not permit
this to happen since it considers both to be semver-compatible. Instead,
by using 0.0.0-alpha.0, we accomplish the same thing without running
afoul of Cargo's semver compatibility rules.

Note that, due to a limitation in `cargo-gnaw`, this commit also
manually edits `third_party/rust_crates/BUILD.gn` slightly. In
particular, `cargo-gnaw` spuriously generates the following:

  group("zerocopy") {
    public_deps = [ ":zerocopy-v0_0_0-alpha_0" ]
  }

  group("zerocopy") {
    public_deps = [ ":zerocopy-v0_8_0-alpha_5" ]
  }

This commit manually edits this to:

  group("zerocopy_0_0_0_alpha_0") {
    public_deps = [ ":zerocopy-v0_0_0-alpha_0" ]
  }

  group("zerocopy") {
    public_deps = [ ":zerocopy-v0_8_0-alpha_5" ]
  }

This is slightly brittle, but future invocations of `fx
update-rustc-third-party` will fail in a fairly obvious way without this
change, and this transition should only last a few days. We also leave a
comment in place so that `git diff` will make it clear what the issue
is.

Bug: 367420574
Change-Id: Ib3d668ee70ee5d0d275e76b7bf84805035010b7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120194
Reviewed-by: Julia Ryan <pineapple@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Fuchsia-Auto-Submit: Joshua Liebow-Feeser <joshlf@google.com>