[kernel] Account for save_state_list during CPU offlining
The MigrateUnpinnedThreads function, which moves threads off a CPU as it
goes offline, did not correctly account for the save_state_list. This
fixes the issue by migrating all of the threads in that list along with
all of the threads in the RunQueue.

Bug: 362353680
Change-Id: I5d0d0ef9158fd684aa13ee543e5ee4515a6cd880
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121248
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>