[ffx] Use absolute paths in isolate dirs (reland)
When the daemon is autostarted, there is no guarantee it starts from
the same directory as the invoking CLI. This means that if an isolate
dir is passed in as a relative path (e.g. `ffx --isolate-dir foo ...`)
the daemon may end up using a different isolate dir. The user-visible
behavior is that CLI cannot make a connection to the daemon (because the
ascendd socket ends up in a different location).

To fix this, when a isolate dir is specified as a relative path, we
turn it into an absolute path. (Note: we cannot fully "canonicalize"
it, resolving sym links, because that doesn't play well with
Google3 build environments.)

Fixed: 362704093
Change-Id: Iec11eab3fb4645a62f7e8277a6f974962ddf813a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1117582
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>