[wlanix] Add integration tests for iface name
This change includes the new WifiStaIfaceBaseTestClass that creates an
iface for use by all tests in the subclass. The sole test in
IfaceCorrectnessAndConsistencyTest checks that the name of the iface
reported by WifiChip is consistent with the name the iface reports for
itself.

Various wlanix TODO bugs are also added to wlanix code.

Bug: 367836470
Test: fx test --e2e --output vim3_vg_wlanix
Change-Id: I01a7659bb57fcc75b62a03767719fd152b21b8c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118695
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Sam Balana <sbalana@google.com>