[ffx] SSH parsing does not interpret debugging as err.
In doing some testing, I was able to get an empty error message by
attempting to connect to a nonexistent device. It was reporting that
using an address was an error, because the line was not something the
parser would catch, given it started with `debug2:`.

Bug: 366238654
Bug: 362976369
Bug: 365638219
Bug: 357958365
Bug: 361140479
Change-Id: Icde48a8ffee8e771d8c10ec8af5459190b6af907
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121661
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>